move test-only method to tests
